added company reference to connection

+++ src/RoutePlanning.Domain/Locations/Services/IShortestDistanceService.cs
@@ -4,5 +4,5 @@ namespace RoutePlanning.Domain.Locations.Services;
 
 public interface IShortestDistanceService : IDomainService
 {
-    int CalculateShortestDistance(Location source, Location target);
+    double CalculateShortestDistance(Location source, Location target);
 }
